+++ packages/medical-domain/src/domains/models/statistics/carbs-statistics.model.ts
@@ -42,4 +42,9 @@ export interface RescueCarbsAveragePerRange {
   rescueCarbsOverrideAverage: number
 }
 
+export interface RecommendedRescueCarbsStatistics {
+  averageValue: number
+  averageOverride: number
+}
+
 export type RescueCarbsAverageStatistics = Map<HoursRange, RescueCarbsAveragePerRange>